revert extraction of user posting from signup function
