+++ src/main/webapp/user.js
@@ -14,38 +14,15 @@ function getUser(user) { //eslint-disable-line
   const fetchRequest = fetch(makeRelativeUrlAbsolute('/user?githubID=' + user));
 
   const errorFormattedFetchRequest = standardizeFetchErrors(
-      fetchRequest,
-      'Failed to communicate with the server. Please try again later.',
-      'An error occcured while retrieving this account.' +
+    fetchRequest,
+    'Failed to communicate with the server. Please try again later.',
+    'An error occcured while retrieving this account.' +
     ' Please try again later.');
 
   return errorFormattedFetchRequest.then((response) => response.json())
-      .then((user) => {
-        return user;
-      });
+    .then((user) => {
+      return user;
+    });
 }
 
-/**
- * Posts a user to the servlet
- * @param {URLSearchParams} userParams parameters with the users information
- * @return {Promise} response from '/user' POST
- */
-function postUser(userParams) {
-  const fetchRequest = fetch(makeRelativeUrlAbsolute('/user'), {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/x-www-form-urlencoded',
-    },
-    body: userParams,
-  });
-
-  const errorFormattedFetchRequest = standardizeFetchErrors(
-      fetchRequest,
-      'Failed to communicate with the server. Please try again later.',
-      'An error occcured while creating your account.' +
-    ' Please try again later.');
-
-  return errorFormattedFetchRequest;
-}
-
-export {getUser, postUser};
+export { getUser };
